fix:Check if the docker component exists

+++ scripts/stop.sh
@@ -2,5 +2,13 @@
 
 function stop_node_scheduler()
 {
-  docker stop cess-boot && docker rm cess-boot
+  echo "Check cess-boot container status"
+  docker ps -a|grep cess-boot
+  if [ $? -eq  0 ]; then
+    echo "Start stopping the cess-boot container"
+    docker stop cess-boot && docker rm cess-boot
+    echo "Successfully stopped the cess-boot container"
+  else
+    echo "cess-boot container has stopped"
+  fi
 }
